Abstract the notification adding and auto-removing into the store itself
